search+libnotmuch: use the FFI to perform a search
An experimental version of the search interface that uses libnotmuch.
This is pretty slow since there are quite a few collecting loops. I'm
guessing the iteration should be deferred until the rendering, maybe
with some sort of stream?

There's also a very idiotic port of the notmuch_time_relative_date
function, since the shell command gives dates in that format instead of
raw. Should be refactored to be more lispy.

One must replace the notmuch-search call in nyxtmuch-render-search with
        (libnotmuch-search search-string "/path/to/db")
to be able to use the FFI version