search: use a new source supporting 'tag:' completion
This implements a somewhat "stateful" source: on construction, it
fetches all notmuch tags from the database, and presents those as
suggestions when the last word starts with "tag:". Otherwise, the
suggestions come from the history.

Also, this introduces two additional commands that pre-fill the prompt
input, `nyxtmuch-search-tag` and `nyxtmuch-clone-search`.

As of nyxt 30a4e02bc86, the pre-filled prompt comes selected, which is
less than ideal. A solution for
https://github.com/atlas-engineer/nyxt/issues/1578 would be nice, and my
own selfish take on it is:

```
    @@ -329,9 +329,18 @@ This does not redraw the whole prompt buffer, unlike `prompt-render'."
                        (get-element-by-id "input")
                        (focus)))))

    +(defun prompt-render-unselect (prompt-buffer)
    +  (ffi-buffer-evaluate-javascript-async
    +   prompt-buffer
    +   (ps:ps (ps:chain document
    +                    (get-element-by-id "input")
    +                    (set-selection-range 999 999)))))
    +
    +
    (defmethod prompt-render ((prompt-buffer prompt-buffer)) ; TODO: Merge into `show-prompt-buffer'?
    (prompt-render-skeleton prompt-buffer)
    (prompt-render-focus prompt-buffer)
    +  (prompt-render-unselect prompt-buffer)
    (prompt-render-suggestions prompt-buffer))

    (defun watch-prompt (prompt-buffer)
```